Renamed GetInvolved folder & files to GetInvolvedScreen and moved folder & files from src to src/screens
