+++ client/src/screens/GetInvolvedScreen/ContactForm/ContactForm.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import "./style.scss";
+import "./ContactForm.scss";
 
 function ContactForm() {
     // we'll have to change the url in action in production
